:bug: Object's 'draw' and 'erase' methods got bug fixes
New comments added in Object.py to address new insights on how
the 'draw' and 'erase' method actually work.

Maybe the way the Object class was designed to draw onto
the screen is not the best. Allowing the user to choose
when to add the 'screen' attribute might not be safe. The
'screen' attr should **always** be set right before the
'draw' method gets called. This would make the code less
bug-prone.