:construction: new fonts added, state pattern (design patter) in use for OptionsBox and more
The game scene now uses the state pattern to control how some of its elements (ex: 'OptionsBox') change whenever an event erupts

The idea is to change the current state of some objects, making them react to changes and adapt their behavior depending on what phase the scene is

The battle scene was thought to be divided in at least 4 states/phases:

1. IDLE State: the player chooses the hero he/she wants to use
2. Action Selection: the player chooses the type of action to perform (attack, defend)
3. Ability Selection: the player chooses the ability to use
4. The battle takes place and all characters have their HP/MP/Stamina bars updated

All these changes are up to revision

Also, new fonts were added. Using new fonts in this project is relatively simple.

The only thing needed to use new fonts is to use their foldername when initializing the 'FontFamily' class.