:construction: battle scene events under development
the 'SelectOption' gameevent needs to get a reference to both the 'FightingCharacter' and the action (atk/defend) chosen by the player.
After that, the scene should draw the options available for that hero.

This commit is about to happen before a new 'test' branch is created, in which the 'Object' class will be tested
with a simpler/better design. The scene's design is also going to change a little bit, with a new idea coming up to
subdivide a scene in states, each one containing its own keyboard/mouse event handler and other specifics.

These ideas are attempts to reduce this project's complexity and improve maintainability. Also, many bugs are
thought to be solved if these changes take place in the future.