:construction: Menu scene updated to work with the redesign of the 'Object' class
The menu's positioning file controls the dimensions and positions of more stuff now

GameEvent/Keyboardhandler/Mousehandler were removed, because they added too much complexity
to this project, without really paying off

Now each scene is responsible to handle the keyboard and mouse by implementing the 'scene_events' method

A 'Grid' class was created. It's responsible to build a representation of a Grid with a list of coordinates.