:tada: battle scene visuals almost done
next steps:

* implement the battle itself
* the player needs to choose what enemy to attack (in case the action
  being performed is an attack)
* all choices have to be stored somehow, so they can be later used to
  compute the round

I'm not really proud of the game so far, maybe it's a bit too
convoluted.

I came to the conclusion that this game depends a lot on how the Object
class is implemented. Even though it went just recently through a
redesign, which was in fact needed, i think the addons system could be
improved once more.

Maybe the Object class as it is today could be split into two classes:
One for an Object without addons and another one for an Object with
addons. That's just an idea i just had, but the point is that something
could be done to make using addons less cumbersome.