:construction: battle's scene box under development
The Box Class has been redesigned to become less complicated, so the use of
the state pattern has been abandoned.

Instead, it now offers methods to paint all phases related to choosing a
hero, choosing an action, choosing an ability and choosing a target.

Also, a custom event called ANIMATE has been created. This event makes
animating stuff on the screen less cumbersome, as it's more reliable than
counting NOEVENT's and reacting to that.

Individual fighters were created and the Combat system is
underdevelopment. For now, it's already possible to make one hero attack another
one, with HP, MP and Stamina bars being correctly updated. This
functionality will probably be toggled in the future, when it get's
integrated with the Box's class.

Wrapping it up, a new class called Combat was created to manage all
turns of the Battle scene. In order to work, it's necessary to ensure
that the Box Class works as expected, since it will be responsible of:
	* collecting user input
	* providing that input to the Combat Class
Besides this, the StatusBox still needs to be created, but this is not really a
priority right now.